# perf(marshal): Skip unnecessary assert.details calls