--- a/packages/marshal/src/helpers/internal-types.js
+++ b/packages/marshal/src/helpers/internal-types.js
@@ -23,7 +23,7 @@ export {};
  *
  * @property {PassStyle} styleName
  *
- * @property {(candidate: any, check: Checker) => boolean} canBeValid
+ * @property {(candidate: any, check?: Checker) => boolean} canBeValid
  * If `canBeValid` returns true, then the candidate would
  * definitely not be valid for any of the other helpers.
  * `assertValid` still needs to be called to determine if it