Ref 11d9f976944b4faaea06037252cfd54e3dd91c37

A modest but significant speedup, especially for big invalid input:

<details><summary>ad hoc analysis</summary>

```sh
$ node --input-type=module -e '
  import Benchmark from "benchmark";
  import "../init/index.js";
  import { makeTagged } from "./src/makeTagged.js";
  import { makeMarshal } from "./src/marshal.js";
  const { serialize } = makeMarshal(x => x, x => x, { errorTagging: "off", serializeBodyFormat: "smallcaps" });
  const rProto = Object.defineProperties({}, { [Symbol.for("passStyle")]: { value: "remotable" }, [Symbol.toStringTag]: { value: "Remotable" } });
  const inputs = [
    undefined,
    "foo",
    true,
    33,
    33n,
    Symbol.for("foo"),
    Symbol.iterator,
    null,
    Promise.resolve(null),
    [3, 4],
    { foo: 3 },
    { then: "non-function then ok" },
    makeTagged("unknown", undefined),
    Error("ok"),
  ].map(v => harden(v));
  const badInputs = [
    Symbol("unique"),
    {},
    harden(Object.setPrototypeOf(Promise.resolve(), { __proto__: Promise.prototype })),
    harden(Object.assign(Promise.resolve(), { extra: "unexpected own property" })),
    harden(Object.defineProperties(Promise.resolve(), { then: { value: () => "bad then" } })),
    harden({ then: () => "thenable" }),
  ];
  let result;
  (new Benchmark.Suite())
    .add("valid inputs", () => {
      for (const v of inputs) result = serialize(v);
    })
    .add("invalid inputs", () => {
      let caught;
      for (const v of badInputs) {
        caught = undefined;
        try {
          result = serialize(v);
        } catch (err) {
          caught = err;
        } finally {
          if (!caught) throw new Error(`unexpected acceptance: ${v}`);
        }
      }
    })
    .add("big input", () => {
      const input = Array(999).fill().map(() => [...inputs]).concat(Object.create(rProto));
      result = serialize(harden(input));
    })
    .add("big invalid input", () => {
      const input = Array(999).fill().map(() => [...inputs]).concat(Object.create(rProto, { own: {} }));
      let caught;
      caught = undefined;
      try {
        result = serialize(harden(input));
      } catch (err) {
        caught = err;
      }
      if (!caught) throw new Error(`unexpected acceptance`);
    })
    .on("cycle", evt => console.log(String(evt.target)))
    .on("complete", () => result)
    .run({ async: true })
'
```

</details>

before:
```
valid inputs x 5,179 ops/sec ±1.26% (87 runs sampled)
invalid inputs x 2,507 ops/sec ±1.30% (83 runs sampled)
big input x 14.41 ops/sec ±2.08% (37 runs sampled)
big invalid input x 49.72 ops/sec ±3.46% (65 runs sampled)
```
after:
```
valid inputs x 5,908 ops/sec ±1.18% (88 runs sampled)
invalid inputs x 2,565 ops/sec ±1.13% (90 runs sampled)
big input x 14.85 ops/sec ±2.01% (40 runs sampled)
big invalid input x 59.31 ops/sec ±1.16% (61 runs sampled)
```