@gibson042 2022-10-13
> I am confused about the title of this PR. I wouldn't call what you're avoiding "serialization"

Serialization seems to be what dominates the cost of ``` X`…` ```, but it's true that I'm actually avoiding the latter rather than the former. Rephrasing would be welcome.